Reorganize into vm/ (internals) and src/ (view)
Trying to represent the entire machine via React hooks/callbacks was
too much of a mind-bender for me.  The functional paradigm is pulling
against how my brain works, but it was a fun exercise for a bit.

New model is that vm/ is an imperative model of the machine, and then
all the files in src/ (maybe should move to view/) are the React
components that show contents and allow interaction.  The parent
React component is Vm.tsx that holds the imperative model as a state
variable and then passes values as props to all the UI components.

There's a hack in Vm.tsx, I have to fake out useState() to convince React
 that state has changed (and re-render required).  This is because we
update the imperative model imperatively (not using "pure functions",
there is internal state and we change it :-) ).  I do this by updating
a timeStep counter that is otherwise unused.  Yuck.