Record lineNumber when parsing and set mode in "neg" instruction
The "neg" instruction is the only one that takes no operands.  It would
be annoying to rewrite everything to handle a possibly-not-present "mode".
If I set the mode to Immediate, then all the logic works, and "immediate"
isn't wrong (there is a K, it's just unused and set to 0 by assemblers).