Generate parser to not depend on fs module
Problem: When trying to "yarn build", an error is reported:

Solution: This is because the generated parser includes a sample main
function (in case you invoke it directly, like "node ebpf.js"), that
just parses a file that you pass as an argument.  This could be
convenient, but it messes up "yarn build"/webpack because the react web
app can't have "require('fs')" in it (there's no filesystem in the
browser).

We're not using this anyway, so what we need to do is pass our own
"moduleMain" option to jison, which will tell it to use our function
instead of its own (that uses fs).  This is somewhat difficult because
there's no way to pass this to the jison cli command, and the parameter
is hidden a couple layers deep in the CLI code, so we have to copy/paste
a good deal, but I didn't see a better option.

All the code needed to invoke the jison generator with our own main is
self-contained in tools/generateParser.js.

Update package.json to invoke this every time we build.