Reorganize, use emscripten MODULARIZE
Emscripten really doesn't want you to pass it a WebAssembly.Memory.
It wants to make its own (it uses some of the heap for its own
initialization purposes and globals, plus your globals, also some
sentinel values).

It's much easier to just take the Emscripten module that it gives you
back and ask it for the heap.  However, couple of headaches:

- Emscripten initializiation is asynchronous.  It involves fetch()ing
WASM and then compiling it (browser probably chooses to do this in a
separate thread).  So we need to do the VM initialization as a Promise.
Vm.tsx is now a wrapper around the VM init - until it's initialized,
it just renders "Initialzing..." and is waiting for the promise.

This "global initialization pause" simplifies things a lot because all
the internal src/vm/ code doesn't have to handle "I'm here but don't
have memory yet"

- We have to take the module.HEAP8.buffer back out and wrap our buffers
around it.