Adjust to use eBPF, successfully load Hello World from BCC
eBPF has a different machine code than BPF.  I guess I thought for some
reason it was an additive extension of BPF (as in, all BPF programs are
valid eBPF programs).  Not quite.

I convinced bcc to emit the bytecode it is loading into the kernel, and
then loaded it in.  It loads in the VM, especially if I have the
external function calls defined (I think it's calling a printf
function).