Single-step through UBPF VM from the GUI
This is the first commit where you can single-step through the UBPF VM
using the GUI.  You can run the entire HelloWorld.cc example program,
and if you check the Javascript console in your browser, you'll see:

    ebpf_trace_printk():  Hello, World! Here I did a sys_clone call!

All the heap and stack memory and CPU state is now reworked to point at
the memory in the UBPF VM (via Uint*Array views on what we get back from
various _ebpfvm_get_stack() instructions and so on that return pointers
into the WASM heap).  They don't construct or hold any of the memory
themselves.

The "step" button calls _ebpfvm_step() which runs one step of the VM.

There's a disassembler in src/vm/instructions.tsx that disassembles the
loaded instructions in the VM.

Function call 6 (ebpf_trace_printk) is overridden: it calls snprintf()
in C and then uses emscripten to call console.log.  This should probably
actually append the string to some ring buffer or something so that we
can directly show it to the user.

There's a lot of missing safety checks along with restartability.  Once
you're done/if you break it, you have to just click "Reload".  But it
works!