Create automated build
Problem: Before this commit, building was a random set of "npm run" /
emcc / sed and copy commands in my shell history, and was not
repeatable.

Solution: Create an automated build.  We need to invoke emcc to compile
the UBPF VM into WASM, and we need to invoke npm to compile all the
typescript and invoke webpack.  Classic GNU make is a reasonable choice
for the overarching build system I think, so I made a Makefile.

This can also be built into a docker container using "docker build",
which does the whole build and then stuffs the build output
(CSS/JS/HTML/WASM) into an nginx container.

The Dockerfile steps are ordered to try to make caching reasonable,
especially for the larger Emscripten SDK and node_modules/ .  If a CI
system is added, it would be nice to cache the yarn cache there so we
don't have to re-download it.