Add two anycasts to disable TS warnings on extra safety checks
Problem: Typescript warns TS2774 on these lines because according
to typescript's checker, these values can never be null, and we probably
mean to call the function instead of checking if the ref is null.

Solution:  This is usually very helpful, but in this case we really are
trying to do an extra check before calling.  We're choosing a defensive
posture for these interface functions.  To prevent the warning, cast
them to "any" (telling typescript "these could actually be null") before
doing the null-check.