Add separatorByte to hex editor
Problem: The hex editor bytes kind of look like a sea of identical bytes
(with a little bit of coloring difference between each byte).  It would
be nice to separate them into words.

Solution: Add concept of separatorByte all over.  It is hardcoded to
every eighth column.  To compute the size for the "measure row" stuff,
we compute the difference in width between a separator byte and a normal
byte.