Add scrollToByte to react-hex-editor
Problem: react-hex-editor exposes scrollTo and scrollToItem methods to
scroll a particular element into view.  However, these are row-based
(they pass through to react-window, which manages the rolling window of
"rows of bytes").  It's difficult to know which row a particular byte is
in from a parent class, because we're using AutosizeHexEditor, so the
number of bytes per row varies as the user adjusts their browser window.

Solution: react-hex-editor's HexEditor class does the scrolling, and it
knows how many columns per row (because it's inside the autosizer).  It
is the right place to do this calculation; the outer class should just
say which byte it wants highlighted.

Add scrollToByte(), which is like scrollToItem() but first computes
which row the referenced byte is in.