Rename eBPF assembler parser and commonize build
Problem: The parser for the eBPF assembler was named "ebpf" which isn't
very clarifying.  It's built when you run various npm steps (like "npm
build"), which isn't great in the current workflow that is Makefile
controlled (it'd be better to not have various build steps scattered all
over a bunch of different places).

Solution: Rename it "ebpf-assembler" and teach the Makefile how to build
it.  We still have to invoke "npm build-assembler" to make sure we get
the right node environment.