Update assembler for eBPF instead of old BPF
Problem: Assembler was for the old BPF syntax.

Solution: Update for eBPF, nearly a re-write of the assembler jison
grammar.  Some notes:

- Remove the addressing mode constants in src/vm/parser/consts.ts.
These aren't helpful outside of the assembler (and even there, not very
helpful) and are not reflected in the resulting instructions.  Use
a consistent, smaller set of consts from src/vm/consts.tsx everywhere.

- Don't emit "%#1234" in the disassembler anymore.  Only the ubpf
disassembler does it, I don't find it in any examples anywhere.  The
ubpf assembler doesn't even allow it.  Just drop it.  (I think if you
ship an assembler and a disassembler, a good practice is that your assembler
can assemble anything your disassembler emits).

- Fix/write a whole bunch of unit tests for the parser.

This doesn't do strong validation; that will be a separate step.  For
instance, the parser will accept "lddw r47, 1234" even though there is
no r47 in eBPF.  Also, negative branch offsets are allowed too (though
eBPF is beginning to allow this in limited circumstances).