Add encoders for load, store, ALU and ALU32 instructions, and assembler
Problem: None of the old emitter/assembler code works for eBPF.

Solution: Rewrite for eBPF.

Fix a bug in the parser, I wasn't allowing negative immediates for ALU
operations ("add r10, -48").  Add the operands_reg_dir_imm class to
account for this.

Fix the byte order in the packer to match uBPF.

Remove "mode" from everywhere.  This is encoded in the opcode plus
source/dest.  I was worried about this ambiguity:

  If source is 0 and imm is 0, is this a mov from r0, or the
  immediate 0?

but this turns out to not be a problem.  The parser differentiates
between source "r0" and source "" (i.e. source is not a register).
Eliminating mode removes ambiguity (some ops don't have a mode) and
reduces the surface area for me to mess up types out of the parser
(because Typescript doesn't understand the parser, so it doesn't verify
that I am properly building ParsedInstructions).

Also add a "test:debug" target so that we can attach devtools to Jest.

The symbol/label resolver is still not really operable; we call it but
it doesn't do much yet.  Also jumps, endian ops, ldabs and ldindx are
not implemented.