Allow empty CPU instruction for CPU state
Problem: The CPU state cannot handle a null CPU instruction, which might
happen after we execute the last instruction in the program (the PC
points at nothing).

Solution: Before we just probably had zeros there and it worked but this
is not ideal.  We should indicate to the user that we have no idea what
the CPU would do if it hypothetically tried to continue executing.
Let the CPU state react component accept "Instruction | null" and render
"-" for all the instruction-related values.