Update UI to use assembled source rather than bytecode
Problem: Before this commit, we were storing the bytecode for hello
world, and then disassembling it to get the source.  This worked but it
meant we could never really edit a program (unless we wanted to edit the
machine code, which we don't).

Solution: Now that the assembler works, we instead assemble
HELLOWORLD_SOURCE and use that everywhere.  That changes a bunch of
interfaces.  The most significant change is that in a Program, the
"lddw" instruction is one statement, and one instruction (that is 16
bytes long).  Before, we turned it into two statements (one was "lddw"
and the other was a blank), each with one instruction that was 8 bytes
long.

Now that instructions can be variable length, there are a few places
where we were converting from program counter to instruction index that
don't work.  We need to use Program.getInstructionAtProgramCounter()
instead.  Someday we should probably cache things like this in Program.

After this, I believe everything works the same in the UI, but we are no
longer relying on HELLOWORLD_HEXBYTECODE.