Allow editing code
Problem: No way for users to edit the code of the BPF program.

Solution: Use codemirror to allow them to edit.  We also now need to
handle their edited program, so we assemble it and then load it into the
VM, and update the read-only view.

- Programs might grow in length beyond when we first inited the VM, so
we allocate a fixed-size buffer (MAX_PROGRAM_SIZE) for instructions, and
then make sure every program we assemble fits inside of it before
loading it (we return an error to the user if it is too big).
MAX_PROGRAM_SIZE is currently big enough (16*512) for 512 "lddw"
instructions, or any program up to 1024 instructions as long as it has
no "lddw" (lddw is the one instruction that is 16 bytes, the rest are
8).

- Initializing the VM is more complicated because we may change the
program while it is running.  There were also some React warnings about
lieing about hook dependencies - this is apparently a recommended
practice in the special case that you are trying to do one-time
initialization (like componentDidMount() in the old days) but it's
complicated and "code smelly" to do it this way.  I created a
VmInitializer class whose job it is to do the one-time initialization
and then pass that internal vm down to the child, the "true" VM wrapper
GUI component.

- Add editing to the program listing.  The program listing component has
four states:
  - Loaded: Read-only view of the currently loaded program.
  - Editing: Read-write, user is typing into CodeMirror.  The last
    assembler error they encountered may be in a box below the program.
  - Assembling: User was in Editing and clicked "Build".  This state is
    transient - in the future, building may be asynchronous so we may
    need to stay in it, but for now, we then immediately (synchronously)
    leave this state and go to either "Loaded" (successfully assembled),
    or HasError (had an error assembling).
  - HasError: The same as Editing, except we have an extra "Error"
indicator right near the "Build" button; this is to hint the user that
they have an error and that's why they're still in the editor (because
this is right next to the "Build" button they just clicked; we're
providing feedback really close to where they're looking).  As soon as
the user types anything,  we move back to Editing (and the actual long
error message still remains).

- Minor, but make the error listings use 'white-space: pre-line' CSS
attribute, which makes them honor '\n's in the text (without making us
turn them into <br/>s)