Render multiline instructions (comments or blanks) properly
Problem: Programs are allowed to have blank lines and comment lines in
them, but if they do, the program display is incorrect.  For instance,
this is a valid program:

    // foo

    mov r1, 40

Solution: Remove the assumption that a statement is exactly one line.
It may be more than one line.  For the purposes of display and
highlighting statements, we "pull in" any preceding comments into the
statement that follows; for example the above program is exactly one
statement, and the statement source code is "// foo\n\nmov r1, 40".

This required fixing a bug in the parser: we said the "NL" token was at
least one "\n".  We should make it exactly one "\n", so that if you have
multiple "\n"s in a row (i.e. any totally blank lines), the parser
treats them each as an independent line, and thus does
"yy.currentLine++" for each line (rather than just once for the greedy
block of "\n"s).  We also now have to teach the parser about
"empty_line".

Add a unit test for this case.