Display the entry point for a program
Problem: We show the source code for a program but this doesn't give any
context for what started the program (the entry point / event).

Solution: The VM doesn't really care so this is basically just a comment
(maybe it would be more powerful in the future - if the entrypoint is a
network filter we could make sure we're giving a packet, and enable the
extra classic BPF modes in the VM).  If there is a special comment at
the beginning of the program describing the entry point, render it.

Add tests.