Fix offset handling in disassembler
Problem: The disassembler would incorrectly disassemble jumps, giving
them wildly incorrect jump offsets.

Solution: The jump offset was actually where the instruction appeared
in the program, in bytes, instead of the offset actually encoded in
the instruction.  This is because the code incorrectly used `offset`
(where the instruction appears in the Uint8Array we're passed)
instead of `off` (the offset encoded in the jump instruction).

Also remove superfluous arguments like "offset" from the disassemble
method.  When I wrote it, I assumed we might call it on a program
somewhere in the middle of memory.  Now we always call it on an entire
Uint8Array (which may itself be a view of another Uint8Array/bigger
chunk of memory).  I still use offset for the inner
disassembleInstruction() to avoid constructing new Uint8Arrays for
every instruction (plus it's hard to manage - some instructions are
lddw's that are 16 bytes, so properly slicing the raw byte array
requires essentially disassembling it first to know where the
instruction boundaries are).