Add entrypoint so we can define program entrypoints
entrypoints are the initial state when the eBPF program starts.
For example, for the sched_fork program, the entrypoint creates
the relevant task_struct in memory (by loading it from task_struct.bin),
as well as the context.  Next, it updates the pointer inside of
context to point at where task_struct ended up.

Note that there are many other pointers in context and task_struct that
are *not* corrected.  How much patching and modification you want to do
is entrypoint and scenario-dependent.