Refactor VM memory interface
Problem: The existing VM memory interface required converting between
absolute and relative addresses a lot, for example a pointer in a
function call (stored in a register) would be an absolute memory
address.  We'd have to either:

a) decide if that address was in the heap or the stack, then subtract
the heap/stack starting offset from the address (to get the relative
address) and then maybe create new BigUint64Array views of that data
to pull it out, or

b) Dig inside of the heap/stack to get the true buffer representing
all of WASM emscripten memory, so we could use the absolute address
somewhat directly (maybe having to create a BigUint64Array view).

Solution: Simplify into one Memory class, which holds a few views
of the same underlying buffer:

 - "all" (and all64, all32): all memory, using absolute addresses.
 - "stack": a view of the stack inside "all", relative.
 - "heap": a view of the heap inside "all", relative.

users can use whichever view is most convenient for them.