Render the size of the hot address
Problem: Our hot address code would scroll the heap/stack view to
the right address but it wasn't clear exactly where in the view the
correct address (bytes being used) are.  We could always highlight
8 bytes starting at the starting address but that isn't correct
because load/stores of 1, 2, 4, and 8 bytes are allowed.

Solution: Add hot_address_size in the ubpf VM, to track the size of
the last read/write we did.  Update Vm and Memory to use it.