Make WindowedProgram use react-virtual for table overflow
This is how the HexEditor works.  However, HexEditor does a lot of work
to figure out the size of a row (rendering one row invisibly just to
measure).  It's fragile and I'm not an expert at that kind of thing so
I'm looking at react-table which presents a simpler API (less for me to
mess up CSS/rendering-wise).