Rewrite code view to use a virtual react-table
Problem: The read-only code view could be very large, and
it should probably remain between memory and CPU.  This
means users may have to scroll past a big program listing often,
when they probably just want a few lines of context.

Solution: Use a virtual table - we only show a few lines that
we're focused on and hide the rest of the lines above/below, and
then we generate them when we scroll them into view.

Still some work to be done after this commit:
1. We are recording whether an address is active but not doing
any thing with it yet.  It'd probably be best to see if we can
use MUI toolkit so it looks consistent with other active/highlighted
items.

2. The cell width is fixed and doesn't look consistent with MUI.

3. Don't have a way to expand/contract the program listing; you get
a fixed window of lines.