Build WindowedProgram using MUI and avoid jitter
We use MUI as our basic UI toolkit, so use it here.  react-window
doesn't care much about styling, so it's mainly replacing <tr> with
<TableRow> from MUI and so on.

The size estimator was way off, it looks like if it is off by more
than a factor of 2, then you can get "jitters" (you render forever)
because react-virtual keeps flipping back and forth on how many
overscan elements to render.  Improve the estimator (it doesn't have
to be perfect but if line size changes dramatically, the constants
should be updated).

Render active style: the PC is black (instead of gray) and the row
is highlighted.