Add wide mode
Problem: Fullscreen on wide monitors there's a huge gutter but you
still can't see all the content.

Solution: Add "wide mode", which will use a responsive flexbox grid
allowing two-columns on wide monitors.  It is not the default.  You
activate it by clicking wide mode on the bottom.

If the autodetection turns out to be good enough this could be automatic
(maybe with an opt-out).